Change time series records to unix timestamps
This makes it easier to update existing records and unix timestamps are
a better standard than date time as timezones are uncertain. At least in
the way date time has been used before. It is now consistent with
theusage of the poloniex API. Also added seed and reward logging
capability to help finding a good model.