Implement saving and restoring of models
To allow serving Pythia model in an online environment, the model needs
to be savable so it can be run from a server after training. To achieve
this the tensor-flow graph and its trainable weights need to be saved.
Additionally to ensure the same performance to to allow for a seamless
transition from training to live, the memory of the agent can be saved
and restored as well.