Add PoloniexConnection class
The poloniex connection should handle retries and buffering of chart
data. This class should be used by the live time series to provide the
agent with live chart data. The retries and telemetry calls are
implemented to make sure that the agent gets an uninterrupted stream of
chart data and that there are no gaps between intervals. If there are
gaps errors are raised and the environment will have to react
appropriately.