Add Telemetry class
This class should handle writing telemetry data out to appropriate
files. It also provides a convenience function to get the time stamp of
the last chart data received for a pair, which is used by the poloniex
connection to determine which period interval to query. Also it handles
rotating the lines in the files, so they do not take up too much space
on the server.