Implement poloniex chart data API
The online paper trader agent needs access to current chart data, hence
the poloniex chart data API needed to be implemented. To make sure the
max API usage of 6 calls per second is not exceeded sleep guards have
been added to the API to prevent calls exceeding the limit.

To be on the safe side minimum random delay is ~1/6 seconds and a calls
are counted and after 4 subsequent calls a wait is forced.