Rename FpmEnvironment to FpmBackTestEnvironment
To distinguish the environment used for back-testing, and the one for
online paper trading the FpmEnvironment has been renamed. This name will
now serve as base class name for both new environments and the future
online environment. Main difference should be the reset functions as
online environments can't be reset.