Implement online paper trading service
To allow for live checking the system an online paper trading service
has been implemented. In the future this should be configurable to
become a real live trading system. For now it can be used to setup all
the basic surrounding components like server frontend, system service
setup etc. Additionally it can be used to test the current coin setup.

It turned out that BTC_XEM takes longer than 30min to provide chart
data. It should be considered to be removed from the portfolio.